add 'functions' module
